backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pydantic-settings==2.7.0
sqlalchemy==2.0.36
alembic==1.14.0
httpx==0.28.1
openai==1.58.1
google-genai==0.8.0
openai-whisper==20240930
ffmpeg-python==0.2.0
pytest==8.3.4
pytest-asyncio==0.24.0
